authenticate request

+++ TodoApp/routers/todos.py
@@ -1,21 +1,13 @@
-from datetime import timedelta, datetime
-from typing import Dict
+from typing import Annotated
 from pydantic import BaseModel, Field
 from sqlalchemy.orm import Session
 from fastapi import APIRouter, Depends, HTTPException, status, Path
-from passlib.context import CryptContext
-from fastapi.security import OAuth2PasswordRequestForm, OAuth2PasswordBearer
-from jose import jwt, JWTError
+from models import Todos
 from database import SessionLocal
-from models import Users, Todos
+from .auth import get_current_user
+
 
 router = APIRouter()
-
-SECRET_KEY = '4d35a986f84841e22b96445ca588c30ce4f4409126a6ea8c46aaa8c58dfda00b'
-ALGORITHM = 'HS256'
-
-bcrypt_context = CryptContext(schemes=['bcrypt'], deprecated='auto')
-oauth2_bearer = OAuth2PasswordBearer(tokenUrl='auth/token')
 
 
 def get_db():
@@ -25,53 +17,8 @@ def get_db():
     finally:
         db.close()
 
-
-db_dependency = Depends(get_db)
-
-
-class CreateUserRequest(BaseModel):
-    username: str
-    email: str
-    first_name: str
-    last_name: str
-    password: str
-    role: str
-
-
-class Token(BaseModel):
-    access_token: str
-    token_type: str
-
-
-async def get_current_user(token: str = Depends(oauth2_bearer)) -> Dict[str, int]:
-    try:
-        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
-        username: str = payload.get('sub')
-        user_id: int = payload.get('id')
-        if username is None or user_id is None:
-            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
-                                detail='Could not validate user.')
-        return {'username': username, 'id': user_id}
-    except JWTError:
-        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
-                            detail='Could not validate user.')
-
-
-def authenticate_user(username: str, password: str, db: Session):
-    user = db.query(Users).filter(Users.username == username).first()
-    if not user:
-        return False
-    if not bcrypt_context.verify(password, user.hashed_password):
-        return False
-    return user
-
-
-def create_access_token(username: str, user_id: int, expires_delta: timedelta):
-    encode = {'sub': username, 'id': user_id}
-    expires = datetime.utcnow() + expires_delta
-    encode.update({'exp': expires})
-    return jwt.encode(encode, SECRET_KEY, algorithm=ALGORITHM)
-
+db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 class TodoRequest(BaseModel):
     title: str = Field(min_length=3)
@@ -80,76 +27,43 @@ class TodoRequest(BaseModel):
     complete: bool
 
 
-@router.post("/users", status_code=status.HTTP_201_CREATED)
-async def create_user(
-        create_user_request: CreateUserRequest,
-        db: Session = db_dependency  # پارامتر با مقدار پیش‌فرض باید در انتها باشد
-):
-    create_user_model = Users(
-        email=create_user_request.email,
-        username=create_user_request.username,
-        first_name=create_user_request.first_name,
-        last_name=create_user_request.last_name,
-        role=create_user_request.role,
-        hashed_password=bcrypt_context.hash(create_user_request.password),
-        is_active=True
-    )
-
-    db.add(create_user_model)
-    db.commit()
-    return create_user_model
-
-
-@router.post("/token", response_model=Token)
-async def login_for_access_token(
-        form_data: OAuth2PasswordRequestForm = Depends(),
-        db: Session = db_dependency
-):
-    user = authenticate_user(form_data.username, form_data.password, db)
-    if not user:
-        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
-                            detail='Could not validate user.')
-    token = create_access_token(user.username, user.id, timedelta(minutes=20))
-    return {'access_token': token, 'token_type': 'bearer'}
-
-
-@router.get('/todos', status_code=status.HTTP_200_OK)
-async def read_all(db: Session = db_dependency):
-    return db.query(Todos).all()
+@router.get('/', status_code=status.HTTP_200_OK)
+async def reade_all(user: user_dependency, db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=401, detail='authentication filed')
+    return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
 
 
 @router.get('/todos/{todo_id}', status_code=status.HTTP_200_OK)
-async def read_todo(todo_id: int = Path(gt=0), db: Session = db_dependency):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
-    if todo_model:
-        return todo_model
-    raise HTTPException(status_code=404, detail='Todo not found')
-
-
-@router.post('/todos', status_code=status.HTTP_201_CREATED)
-async def create_todo(
-        user: Dict[str, int] = Depends(get_current_user),
-        db: Session = db_dependency,
-        todo_request: TodoRequest
-):
+async def read_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
     if user is None:
-        raise HTTPException(status_code=401, detail='Authentication Failed')
+        raise HTTPException(status_code=401, detail='authentication filed')
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
 
+    if todo_model is not None:
+        return todo_model
+    raise HTTPException(status_code=404, detail='todo not found')
+
+@router.post('/todo', status_code=status.HTTP_201_CREATED)
+async def create_todo(user: user_dependency, db: db_dependency,
+                      todo_request: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=401, detail='authentication filed')
     todo_model = Todos(**todo_request.dict(), owner_id=user.get('id'))
+
     db.add(todo_model)
     db.commit()
-    return todo_model
 
 
 @router.put("/todos/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(
-        todo_id: int = Path(gt=0),
-        todo_request: TodoRequest,
-        db: Session = db_dependency
-):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def update_todo(user: user_dependency, db: db_dependency, todo_request: TodoRequest, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='authentication filed')
+
+    todo_model = (db.query(Todos).filter(Todos.id == todo_id)\
+                  .filter(Todos.owner_id == user.get('id')).first())
     if todo_model is None:
-        raise HTTPException(status_code=404, detail="Todo id not found")
+        raise HTTPException(status_code=404, detail="todo id not found")
 
     todo_model.title = todo_request.title
     todo_model.description = todo_request.description
@@ -158,17 +72,17 @@ async def update_todo(
 
     db.add(todo_model)
     db.commit()
-    return None
 
 
-@router.delete('/todos/{todo_id}', status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(
-        todo_id: int = Path(gt=0),
-        db: Session = db_dependency
-):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+@router .delete('/todos/{todo_id}', status_code=status.HTTP_204_NO_CONTENT)
+async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='authentication filed')
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+        .filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
-        raise HTTPException(status_code=404, detail='Todo not found')
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+        raise HTTPException(status_code=404, detail='todo not found')
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).delete()
+
     db.commit()
-    return None
